resume-parser-app/requirements.txt
Flask
google-generativeai
PyPDF2
python-dotenv